refactor: 커멘드 분리
'punishment'와 'punishment-tracking-end'명령어 분리
